requirements.txt
openai
telethon
python-dotenv
supabase
gspread
oauth2client
json
requests
